requirements.txt
mariadb
pyfingerprint
rpi.gpio
